requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.32.1
awscrt==0.28.4
Babel==2.9.1
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
bs4==0.0.2
cachetools==6.2.4
certifi==2026.1.4
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.4.4
chevron==0.13.1
click==8.1.8
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.10
exceptiongroup==1.3.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gpg==1.23.2
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
idna==2.10
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
langchain==0.3.27
langchain-aws==0.2.35
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==1.1.1
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
netifaces==0.10.6
numpy==1.26.4
oauthlib==3.0.2
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.20
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.1
pypdf==6.6.0
pypdfium2==5.3.0
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.32.5
requests-toolbelt==1.0.0
rpm==4.16.1.3
rsa==4.9.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
s3transfer==0.16.0
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
smmap==5.0.2
soupsieve==2.8.2
SQLAlchemy==2.0.45
streamlit==1.50.0
support-info==1.0
systemd-python==235
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==1.25.10
uuid_utils==0.13.0
watchdog==6.0.0
wcwidth==0.2.5
yarl==1.22.0
zstandard==0.25.0
